requirements.txt
flask==3.0.0
python-telegram-bot==20.7